+++ pkg/provider/basic.go
@@ -32,21 +32,23 @@ func init() {
 }
 
 type basicOperations struct {
-	endpoint     Endpoint
-	clientID     string
-	clientSecret string
+	vsn             int
+	endpointFactory EndpointFactoryFunc
+	clientID        string
+	clientSecret    string
 }
 
 func (bo *basicOperations) AuthCodeURL(state string, opts ...AuthCodeURLOption) (string, bool) {
-	if bo.endpoint.AuthURL == "" {
-		return "", false
-	}
-
 	o := &AuthCodeURLOptions{}
 	o.ApplyOptions(opts)
 
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+	if endpoint.AuthURL == "" {
+		return "", false
+	}
+
 	cfg := &oauth2.Config{
-		Endpoint:    bo.endpoint.Endpoint,
+		Endpoint:    endpoint.Endpoint,
 		ClientID:    bo.clientID,
 		Scopes:      o.Scopes,
 		RedirectURL: o.RedirectURL,
@@ -56,20 +58,21 @@ func (bo *basicOperations) AuthCodeURL(state string, opts ...AuthCodeURLOption) 
 }
 
 func (bo *basicOperations) DeviceCodeAuth(ctx context.Context, opts ...DeviceCodeAuthOption) (*devicecode.Auth, bool, error) {
-	if bo.endpoint.DeviceURL == "" {
-		return nil, false, nil
-	}
-
 	o := &DeviceCodeAuthOptions{}
 	o.ApplyOptions(opts)
 
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+	if endpoint.DeviceURL == "" {
+		return nil, false, nil
+	}
+
 	cfg := &devicecode.Config{
 		Config: &oauth2.Config{
-			Endpoint: bo.endpoint.Endpoint,
+			Endpoint: endpoint.Endpoint,
 			ClientID: bo.clientID,
 			Scopes:   o.Scopes,
 		},
-		DeviceURL: bo.endpoint.DeviceURL,
+		DeviceURL: endpoint.DeviceURL,
 	}
 
 	auth, err := cfg.DeviceCodeAuth(ctx)
@@ -77,12 +80,17 @@ func (bo *basicOperations) DeviceCodeAuth(ctx context.Context, opts ...DeviceCod
 }
 
 func (bo *basicOperations) DeviceCodeExchange(ctx context.Context, deviceCode string, opts ...DeviceCodeExchangeOption) (*Token, error) {
+	o := &DeviceCodeExchangeOptions{}
+	o.ApplyOptions(opts)
+
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+
 	cfg := &devicecode.Config{
 		Config: &oauth2.Config{
-			Endpoint: bo.endpoint.Endpoint,
+			Endpoint: endpoint.Endpoint,
 			ClientID: bo.clientID,
 		},
-		DeviceURL: bo.endpoint.DeviceURL,
+		DeviceURL: endpoint.DeviceURL,
 	}
 
 	tok, err := cfg.DeviceCodeExchange(ctx, deviceCode)
@@ -99,15 +107,22 @@ func (bo *basicOperations) DeviceCodeExchange(ctx context.Context, deviceCode st
 		return nil, err
 	}
 
-	return &Token{Token: tok}, nil
+	return &Token{
+		Token: tok,
+
+		ProviderVersion: bo.vsn,
+		ProviderOptions: o.ProviderOptions,
+	}, nil
 }
 
 func (bo *basicOperations) AuthCodeExchange(ctx context.Context, code string, opts ...AuthCodeExchangeOption) (*Token, error) {
 	o := &AuthCodeExchangeOptions{}
 	o.ApplyOptions(opts)
 
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+
 	cfg := &oauth2.Config{
-		Endpoint:     bo.endpoint.Endpoint,
+		Endpoint:     endpoint.Endpoint,
 		ClientID:     bo.clientID,
 		ClientSecret: bo.clientSecret,
 		RedirectURL:  o.RedirectURL,
@@ -118,12 +133,23 @@ func (bo *basicOperations) AuthCodeExchange(ctx context.Context, code string, op
 		return nil, semerr.Map(err)
 	}
 
-	return &Token{Token: tok}, nil
+	return &Token{
+		Token: tok,
+
+		ProviderVersion: bo.vsn,
+		ProviderOptions: o.ProviderOptions,
+	}, nil
 }
 
 func (bo *basicOperations) RefreshToken(ctx context.Context, t *Token, opts ...RefreshTokenOption) (*Token, error) {
+	o := &RefreshTokenOptions{}
+	WithProviderOptions(t.ProviderOptions).ApplyToRefreshTokenOptions(o)
+	o.ApplyOptions(opts)
+
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+
 	cfg := &oauth2.Config{
-		Endpoint:     bo.endpoint.Endpoint,
+		Endpoint:     endpoint.Endpoint,
 		ClientID:     bo.clientID,
 		ClientSecret: bo.clientSecret,
 	}
@@ -135,18 +161,25 @@ func (bo *basicOperations) RefreshToken(ctx context.Context, t *Token, opts ...R
 		return nil, semerr.Map(err)
 	}
 
-	return &Token{Token: tok}, nil
+	return &Token{
+		Token: tok,
+
+		ProviderVersion: bo.vsn,
+		ProviderOptions: o.ProviderOptions,
+	}, nil
 }
 
 func (bo *basicOperations) ClientCredentials(ctx context.Context, opts ...ClientCredentialsOption) (*Token, error) {
 	o := &ClientCredentialsOptions{}
 	o.ApplyOptions(opts)
 
+	endpoint := bo.endpointFactory(o.ProviderOptions)
+
 	cc := &clientcredentials.Config{
 		ClientID:       bo.clientID,
 		ClientSecret:   bo.clientSecret,
-		TokenURL:       bo.endpoint.TokenURL,
-		AuthStyle:      bo.endpoint.AuthStyle,
+		TokenURL:       endpoint.TokenURL,
+		AuthStyle:      endpoint.AuthStyle,
 		Scopes:         o.Scopes,
 		EndpointParams: o.EndpointParams,
 	}
@@ -156,12 +189,17 @@ func (bo *basicOperations) ClientCredentials(ctx context.Context, opts ...Client
 		return nil, semerr.Map(err)
 	}
 
-	return &Token{Token: tok}, nil
+	return &Token{
+		Token: tok,
+
+		ProviderVersion: bo.vsn,
+		ProviderOptions: o.ProviderOptions,
+	}, nil
 }
 
 type basic struct {
-	vsn      int
-	endpoint Endpoint
+	vsn             int
+	endpointFactory EndpointFactoryFunc
 }
 
 func (b *basic) Version() int {
@@ -174,9 +212,10 @@ func (b *basic) Public(clientID string) PublicOperations {
 
 func (b *basic) Private(clientID, clientSecret string) PrivateOperations {
 	return &basicOperations{
-		endpoint:     b.endpoint,
-		clientID:     clientID,
-		clientSecret: clientSecret,
+		vsn:             b.vsn,
+		endpointFactory: b.endpointFactory,
+		clientID:        clientID,
+		clientSecret:    clientSecret,
 	}
 }
 
@@ -195,35 +234,46 @@ func BasicFactory(endpoint Endpoint) FactoryFunc {
 		}
 
 		p := &basic{
-			vsn:      vsn,
-			endpoint: endpoint,
+			vsn:             vsn,
+			endpointFactory: StaticEndpointFactory(endpoint),
 		}
 		return p, nil
 	}
 }
 
 func AzureADFactory(ctx context.Context, vsn int, opts map[string]string) (Provider, error) {
-	vsn = selectVersion(vsn, 1)
+	vsn = selectVersion(vsn, 2)
+
+	tenant := opts["tenant"]
 
 	switch vsn {
+	case 2:
 	case 1:
+		if tenant == "" {
+			return nil, &OptionError{Option: "tenant", Cause: fmt.Errorf("tenant is required")}
+		}
 	default:
 		return nil, ErrNoProviderWithVersion
 	}
 
-	tenant := opts["tenant"]
-	if tenant == "" {
-		return nil, &OptionError{Option: "tenant", Cause: fmt.Errorf("tenant is required")}
-	}
-
-	// Upstream function does not escape this name, so we will here.
-	tenant = url.PathEscape(tenant)
-
 	p := &basic{
-		vsn: 1,
-		endpoint: Endpoint{
-			Endpoint:  microsoft.AzureADEndpoint(tenant),
-			DeviceURL: "https://login.microsoftonline.com/" + tenant + "/oauth2/v2.0/devicecode", // https://docs.microsoft.com/en-us/azure/active-directory/develop/v2-oauth2-device-code
+		vsn: vsn,
+		endpointFactory: func(credOpts map[string]string) Endpoint {
+			chosenTenant := "organizations" // https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-v2-protocols#endpoints
+
+			if tenant != "" {
+				chosenTenant = tenant
+			} else if credTenant := credOpts["tenant"]; credTenant != "" {
+				chosenTenant = credTenant
+			}
+
+			// Upstream function does not escape this name, so we will here.
+			chosenTenant = url.PathEscape(chosenTenant)
+
+			return Endpoint{
+				Endpoint:  microsoft.AzureADEndpoint(chosenTenant),
+				DeviceURL: "https://login.microsoftonline.com/" + chosenTenant + "/oauth2/v2.0/devicecode", // https://docs.microsoft.com/en-us/azure/active-directory/develop/v2-oauth2-device-code
+			}
 		},
 	}
 	return p, nil
@@ -276,8 +326,8 @@ func CustomFactory(ctx context.Context, vsn int, opts map[string]string) (Provid
 	}
 
 	p := &basic{
-		vsn:      vsn,
-		endpoint: endpoint,
+		vsn:             vsn,
+		endpointFactory: StaticEndpointFactory(endpoint),
 	}
 	return p, nil
 }